+++ platform/linux/keyboard.c
@@ -20,44 +20,45 @@
 
 #define MODULE_NAME "keyboard"
 
-#define N_KEYBOARD_MODES 4
-#define KB_MODES_LIST   \
-    X_(KB_MODE_FAIL)    \
-    X_(KB_MODE_EVDEV)   \
-    X_(KB_MODE_TTY)     \
-    X_(KB_MODE_X11)     \
+#define N_KEYBOARD_TYPES 4
+#define KB_TYPES_LIST   \
+    X_(KB_TYPE_FAIL)    \
+    X_(KB_TYPE_EVDEV)   \
+    X_(KB_TYPE_TTY)     \
+    X_(KB_TYPE_X11)     \
 
 #define X_(name) name,
-enum keyboard_mode {
-    KB_MODES_LIST
+enum keyboard_type {
+    KB_TYPES_LIST
 };
 #undef X_
 
 #define X_(name) #name,
-static const char *keyboard_mode_strings[N_KEYBOARD_MODES] = {
-    KB_MODES_LIST
+static const char *keyboard_type_strings[N_KEYBOARD_TYPES] = {
+    KB_TYPES_LIST
 };
 #undef X_
 
-static const enum keyboard_mode fallback_modes[N_WINDOW_TYPES][N_KEYBOARD_MODES] = {
+static const enum keyboard_type
+fallback_types[N_WINDOW_TYPES][N_KEYBOARD_TYPES] = {
     [WINDOW_TYPE_FRAMEBUFFER] = {
-        KB_MODE_EVDEV,
-        KB_MODE_TTY,
-        KB_MODE_FAIL
+        KB_TYPE_EVDEV,
+        KB_TYPE_TTY,
+        KB_TYPE_FAIL
     },
     [WINDOW_TYPE_X11] = {
-        KB_MODE_X11,
-        KB_MODE_FAIL
+        KB_TYPE_X11,
+        KB_TYPE_FAIL
     },
     [WINDOW_TYPE_DUMMY] = {
-        KB_MODE_EVDEV,
-        KB_MODE_TTY,
-        KB_MODE_FAIL
+        KB_TYPE_EVDEV,
+        KB_TYPE_TTY,
+        KB_TYPE_FAIL
     },
 };
 
 struct p_keyboard {
-    enum keyboard_mode mode;
+    enum keyboard_type type;
     
     union {
         struct keyboard_evdev evdev;
@@ -74,11 +75,11 @@ struct p_keyboard * p_keyboard_init(struct p_window *win)
     s_assert(kb != NULL, "calloc() failed for struct p_keyboard");
 
     enum window_type win_type;
-#define DEFAULT_WINDOW_FALLBACK_MODE WINDOW_TYPE_FRAMEBUFFER
+#define DEFAULT_WINDOW_FALLBACK_TYPE WINDOW_TYPE_FRAMEBUFFER
     if (win == NULL) {
         s_log_warn("Cannot get window metadata. Assuming window type is %s",
-            window_type_strings[DEFAULT_WINDOW_FALLBACK_MODE]);
-        win_type = DEFAULT_WINDOW_FALLBACK_MODE;
+            window_type_strings[DEFAULT_WINDOW_FALLBACK_TYPE]);
+        win_type = DEFAULT_WINDOW_FALLBACK_TYPE;
     } else {
         win_type = win->type;
     }
@@ -86,36 +87,37 @@ struct p_keyboard * p_keyboard_init(struct p_window *win)
     u32 i = 0;
     do {
         s_log_debug("Attempting keyboard init with mode \"%s\"...",
-            keyboard_mode_strings[ fallback_modes[win_type][i] ]
+            keyboard_type_strings[ fallback_types[win_type][i] ]
         );
-        switch (fallback_modes[win_type][i]) {
-            case KB_MODE_EVDEV:
+        switch (fallback_types[win_type][i]) {
+            case KB_TYPE_EVDEV:
                 if (evdev_keyboard_init(&kb->evdev))
-                    s_log_warn("Failed to set up keyboard using /dev/input");
+                    s_log_warn("Failed to set up keyboard using event devices");
                 else
                     goto keyboard_setup_success;
                 break;
-            case KB_MODE_TTY:
+            case KB_TYPE_TTY:
                 if (tty_keyboard_init(&kb->tty))
-                    s_log_warn("Failed to set up keyboard using tty");
+                    s_log_warn("Failed to set up keyboard using tty stdin");
                 else
                     goto keyboard_setup_success;
                 break;
-            case KB_MODE_X11:
+            case KB_TYPE_X11:
                 if (X11_keyboard_init(&kb->x11, &win->x11))
                     s_log_warn("Failed to set up keyboard with X11");
                 else
                     goto keyboard_setup_success;
                 break;
-            default: case KB_MODE_FAIL:
+            default: case KB_TYPE_FAIL:
                 goto_error("Failed to set up keyboard in any known way");
                 break;
         };
-    } while (++i < N_KEYBOARD_MODES);
+    } while (++i < N_KEYBOARD_TYPES);
 
 keyboard_setup_success:
-    kb->mode = fallback_modes[win_type][i];
-    s_log_info("%s() OK, keyboard is in mode \"%s\"", __func__, keyboard_mode_strings[kb->mode]);
+    kb->type = fallback_types[win_type][i];
+    s_log_info("%s() OK, keyboard type is \"%s\"",
+        __func__, keyboard_type_strings[kb->type]);
 
     return kb;
 
@@ -128,14 +130,14 @@ void p_keyboard_update(struct p_keyboard *kb)
 {
     if (kb == NULL) return;
 
-    switch (kb->mode) {
-        case KB_MODE_TTY:
+    switch (kb->type) {
+        case KB_TYPE_TTY:
             tty_keyboard_update_all_keys(&kb->tty, kb->keys);
             break;
-        case KB_MODE_EVDEV:
+        case KB_TYPE_EVDEV:
             evdev_keyboard_update_all_keys(&kb->evdev, kb->keys);
             break;
-        case KB_MODE_X11:
+        case KB_TYPE_X11:
             X11_keyboard_update_all_keys(&kb->x11, kb->keys);
             break;
         default:
@@ -155,22 +157,23 @@ void p_keyboard_destroy(struct p_keyboard **kb_p)
     if (kb_p == NULL || *kb_p == NULL) return;
     struct p_keyboard *kb = *kb_p;
 
-    s_log_debug("Destroying keyboard (mode \"%s\")...", keyboard_mode_strings[kb->mode]);
-    switch(kb->mode) {
-        case KB_MODE_EVDEV:
+    s_log_debug("Destroying keyboard (mode \"%s\")...",
+        keyboard_type_strings[kb->type]);
+    switch(kb->type) {
+        case KB_TYPE_EVDEV:
             evdev_keyboard_destroy(&kb->evdev);
             break;
-        case KB_MODE_TTY:
+        case KB_TYPE_TTY:
             tty_keyboard_destroy(&kb->tty);
             break;
-        case KB_MODE_X11:
+        case KB_TYPE_X11:
             X11_keyboard_destroy(&kb->x11);
             break;
-        default: case KB_MODE_FAIL:
+        default: case KB_TYPE_FAIL:
             break;
     }
 
     u_nzfree(kb);
 }
 
-#undef KB_MODES_LIST
+#undef KB_TYPES_LIST